pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Kamran <diyeqo@icloud.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
flake8 = "^6.0.0"
pytest = "^7.2.2"
beautifulsoup4 = "^4.12.0"
pytest-cov = "^4.0.0"
coverage = "^7.2.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
validators = "^0.20.0"
psycopg2-binary = "^2.9.6"




[tool.poetry.group.dev.dependencies]
flask-testing = "^0.8.1"
blinker = "^1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
